fix: dynamic server error 수정
